LayoutTests/imported/w3c/web-platform-tests/scroll-to-text-fragment/non-html-documents-json-expected.txt


FAIL Text directive blocked in application/json assert_equals: scrolled to fragment expected false but got true
